src/backend/requirements.txt
pandas
python-pptx==0.6.21
matplotlib
numpy